+++ test/mock/dump.dart
@@ -1,17 +1,25 @@
 import 'package:ms_persist/ms_persist.dart';
 
 class Dumb with Persist<Dumb> {
-  String uuid;
-  String dummy;
-  String title;
+  String? uuid;
+  String? dummy;
+  String? title;
+  int beforeSaveHits = 0;
+  int afterSaveHits = 0;
+  int beforeDeleteHits = 0;
+  int afterDeleteHits = 0;
 
-  Dumb({this.uuid, this.dummy, this.title});
+  Dumb({
+    this.uuid,
+    this.dummy,
+    this.title,
+  });
 
   factory Dumb.fromMap(Map<String, dynamic> map) {
     return new Dumb(
-      uuid: map['uuid'] as String,
-      dummy: map['dummy'] as String,
-      title: map['title'] as String,
+      uuid: map['uuid'] as String?,
+      dummy: map['dummy'] as String?,
+      title: map['title'] as String?,
     );
   }
 
@@ -26,4 +34,24 @@ class Dumb with Persist<Dumb> {
 
   @override
   Dumb buildModel(Map<String, dynamic> map) => Dumb.fromMap(map);
+
+  @override
+  void onBeforeSave(Dumb b, bool update) {
+    beforeSaveHits++;
+  }
+
+  @override
+  void onAfterSave(Dumb b) {
+    afterSaveHits++;
+  }
+
+  @override
+  void onBeforeDelete(Dumb b) {
+    beforeDeleteHits++;
+  }
+
+  @override
+  void onAfterDelete(Dumb b) {
+    afterDeleteHits++;
+  }
 }