+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [2.0.0] - 2024/07/15
 
 * Hooks onBefore(Save|Delete) and onAfter(Save|Delete) are now called in the method `save` and `delete` respectively.
+* Added support for Web and Windows platforms.
+* Added support for null safety.
 
 ## [1.0.0] - 2020/09/24